requirements.txt
fastapi
uvicorn
redis
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
python-multipart